Create a transform class and move all the maths operations in there
